feat(module-complete): implement JWT auth and functionalities based on user role

+++ client/src/app/controllers/delete.controller.js
@@ -5,19 +5,30 @@
     'use strict';
     angular.module('titleflix')
         .controller('deleteController',deleteController);
-    deleteController.$inject = ['titleService','$location','$routeParams','Notification'];
-    function deleteController(titleService,$location,$routeParams,Notification) {
+    deleteController.$inject = ['titleService','$location','$routeParams','Notification','$localStorage','userService'];
+    function deleteController(titleService,$location,$routeParams,Notification,$localStorage,userService) {
         var deleteVm = this;
+        deleteVm.userObj = "";
         deleteVm.deleteTitle = deleteTitle;
         function deleteTitle() {
-            titleService.deleteTitle($routeParams.id)
-                .then(function (deletedTitle) {
-                    deleteVm.deletedTitle = deletedTitle;
-                    Notification.success('Title deleted successfully');
-                    $location.path('/admin/browse');
-                },function (error) {
-                    Notification.error('Title not deleted, try again!!!');
-                });
+
+            deleteVm.userObj =  userService.getUserObj($localStorage.userObject);
+
+            if(ddeleteVm.userObj.role === "admin") {
+
+                titleService.deleteTitle($routeParams.id)
+                    .then(function (deletedTitle) {
+                        deleteVm.deletedTitle = deletedTitle;
+                        $location.path('/admin/browse');
+                        Notification.success('Title deleted successfully');
+                    }, function (error) {
+                        Notification.error('Title not deleted, try again!!!');
+                    });
+            }
+            else
+            {
+                Notification.error('Sorry!! not authorized for this functionlity');
+            }
         };
     };
 })();
